pyproject.toml
[project]
name = "enerlytics"
version = "2.0.0"
description = "High-resolution deterministic climate and energy database for India"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.1.0",
    "pandas>=2.2.0",
    "pyarrow>=18.0.0",
    "scipy>=1.14.0",
    "streamlit>=1.40.0",
    "plotly>=5.24.0",
    "pgeocode>=0.4.1",
    "requests>=2.32.0",
    "pvlib>=0.11.0",
    "reportlab>=4.0.0",
    "matplotlib>=3.9.0",
    "tqdm>=4.66.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
enerlytics = "energy_explore.cli:main"
enerlytics-gui = "energy_explore.app:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = false
